* develop(1251-show-lambda-logs): fix tests
* feat(1251-show-lambda-logs): add lambda logs to UI

* feat(1251-show-lambda-logs): fix lambda details page

* Merge branch 'main' into 1251-show-lambda-logs

* feat(1251-show-lambda-logs): fix lambda details page

* Merge branch 'main' into 1251-show-lambda-logs

* develop(1251-show-lambda-logs): fix lambda details page

* Merge branch 'main' into 1251-show-lambda-logs

* develop(1251-show-lambda-logs): fix lambda details page

* Merge branch 'main' into 1251-show-lambda-logs

* develop(1251-show-lambda-logs): fix tests

* Merge branch 'main' into 1251-show-lambda-logs

+++ src/db/src/memorydb/KMSMemoryDb.cpp
@@ -111,7 +111,7 @@ namespace AwsMock::Database {
 
     long KMSMemoryDb::DeleteAllKeys() {
         boost::mutex::scoped_lock lock(_keyMutex);
-        const long deleted = _keys.size();
+        const long deleted = static_cast<long>(_keys.size());
         _keys.clear();
         log_debug << "All KMS keys deleted, count: " << deleted;
         return deleted;
